+++ src/univariate/continuous/noncentralchisq.jl
@@ -36,7 +36,7 @@ end
 
 NoncentralChisq(ν::T, λ::T, ::NoArgCheck) where {T<:Real} = NoncentralChisq{T}(ν, λ)
 NoncentralChisq(ν::Real, λ::Real) = NoncentralChisq(promote(ν, λ)...)
-NoncentralChisq(ν::Integer, λ::Integer) = NoncentralChisq(Float64(ν), Float64(λ))
+NoncentralChisq(ν::Integer, λ::Integer) = NoncentralChisq(float(ν), float(λ))
 
 @distr_support NoncentralChisq 0.0 Inf
