+++ src/univariate/continuous/noncentralf.jl
@@ -15,7 +15,7 @@ end
 
 NoncentralF(ν1::T, ν2::T, λ::T, ::NoArgCheck) where {T<:Real} = NoncentralF{T}(ν1, ν2, λ)
 NoncentralF(ν1::Real, ν2::Real, λ::Real) = NoncentralF(promote(ν1, ν2, λ)...)
-NoncentralF(ν1::Integer, ν2::Integer, λ::Integer) = NoncentralF(Float64(ν1), Float64(ν2), Float64(λ))
+NoncentralF(ν1::Integer, ν2::Integer, λ::Integer) = NoncentralF(float(ν1), float(ν2), float(λ))
 
 @distr_support NoncentralF 0.0 Inf
