+++ src/univariate/discrete/hypergeometric.jl
@@ -65,9 +65,9 @@ end
 skewness(d::Hypergeometric) = (d.nf-d.ns)*sqrt(d.ns+d.nf-1)*(d.ns+d.nf-2*d.n)/sqrt(d.n*d.ns*d.nf*(d.ns+d.nf-d.n))/(d.ns+d.nf-2)
 
 function kurtosis(d::Hypergeometric)
-    ns = Float64(d.ns)
-    nf = Float64(d.nf)
-    n = Float64(d.n)
+    ns = float(d.ns)
+    nf = float(d.nf)
+    n = float(d.n)
     N = ns + nf
     a = (N-1) * N^2 * (N * (N+1) - 6*ns * (N-ns) - 6*n*(N-n)) + 6*n*ns*(nf)*(N-n)*(5*N-6)
     b = (n*ns*(N-ns) * (N-n)*(N-2)*(N-3))
